+++ web/packages/design/src/ButtonIcon/constants.ts
@@ -1,0 +1,35 @@
+/**
+ * Teleport
+ * Copyright (C) 2025 Gravitational, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+export const buttonSizes = {
+  0: {
+    fontSize: '12px',
+    height: '24px',
+    width: '24px',
+  },
+  1: {
+    fontSize: '16px',
+    height: '32px',
+    width: '32px',
+  },
+  2: {
+    fontSize: '24px',
+    height: '48px',
+    width: '48px',
+  },
+};
